TST: Move remaining ewm tests to window/
